==> 1_lingregression.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

taxiDB = pd.read_csv('taxi_dataset.csv')

taxiDB.head(5)

<dl>
<dt> Описание колонок:
<dd>id - ID поездки </dd>
<dd>vendor_id - ID компании, осуществляющей перевозку </dd>
<dd>pickup_datetime - Таймкод начала поездки</dd>
<dd>dropoff_datetime - Таймкод конца поездки </dd>
<dd>passenger_count - Количество пассажиров </dd>
<dd>pickup_longitude - Долгота точки, в которой началась поездка </dd>
<dd>pickup_latitude - Широта точки, в которой началась поездка </dd>
<dd>dropoff_longitude - Долгота точки, в которой закончилась поездка </dd>
<dd>dropoff_latitude - Широта точки, в которой закончилась поездка </dd>
<dd>store_and_fwd_flag - Yes/No: Была ли информация сохранена в памяти транспортного средства из-за потери соединения с сервером </dd>
</dl>

#создаем колонку "продолжительность поездки"
taxiDB['trip_duration'] = (pd.to_datetime(taxiDB['dropoff_datetime']) - \
                           pd.to_datetime(taxiDB['pickup_datetime'])).dt.total_seconds()

taxiDB.head()

#удаляем dropoff_datetime, чтобы убрать data leakage
taxiDB = taxiDB.drop('dropoff_datetime', axis=1)

taxiDB.shape

taxiDB.set_index('id').to_csv('initial_data.csv')

#приводим к виду 0/1
taxiDB['vendor_id'] = taxiDB['vendor_id'] - 1
taxiDB.head()

#приводим к виду 0/1
taxiDB['store_and_fwd_flag'] = taxiDB['store_and_fwd_flag'].apply(lambda x: 0 if x=='N' else 1)
taxiDB.head()

allLat  = list(taxiDB['pickup_latitude']) + list(taxiDB['dropoff_latitude'])

#медианная широта
medianLat  = sorted(allLat)[int(len(allLat)/2)]

#перевод широты в километры
latMultiplier  = 111.32

taxiDB['pickup_latitude']   = latMultiplier  * (taxiDB['pickup_latitude']   - medianLat)
taxiDB['dropoff_latitude']   = latMultiplier  * (taxiDB['dropoff_latitude']  - medianLat)

#на сколько примерно километров севернее или южнее (в зависимости от знака) точка находится относительно медианной широты
taxiDB.head()

#аналогично для долготы
allLong = list(taxiDB['pickup_longitude']) + list(taxiDB['dropoff_longitude'])

medianLong  = sorted(allLong)[int(len(allLong)/2)]

longMultiplier = np.cos(medianLat*(np.pi/180.0)) * 111.32

taxiDB['pickup_longitude']  = longMultiplier * (taxiDB['pickup_longitude']  - medianLong)
taxiDB['dropoff_longitude']  = longMultiplier * (taxiDB['dropoff_longitude'] - medianLong)

taxiDB.head(5)

#вычисление географического расстояния "distance_km":
taxiDB['long_diff'] = taxiDB['dropoff_longitude'] - taxiDB['pickup_longitude']
taxiDB['lat_diff'] = taxiDB['dropoff_latitude'] - taxiDB['pickup_latitude']

taxiDB['distance_km'] = (taxiDB['long_diff']**2 + taxiDB['lat_diff']**2)**(1/2)

taxiDB = taxiDB.drop(['long_diff', 'lat_diff'], axis=1)

taxiDB.head()

#убираем старые признаки
taxiDB = taxiDB.drop(['pickup_longitude', 'dropoff_longitude',
                      'pickup_latitude', 'dropoff_latitude'], axis=1)

taxiDB.head()

taxiDB['passenger_count'].value_counts()

#mean target encoding количества пассажиров
taxiDB['passenger_count'] = taxiDB['passenger_count'].map(taxiDB
                                                          .groupby('passenger_count')['trip_duration']
                                                          .mean())

taxiDB.head()

taxiDB = taxiDB.set_index('id')

taxiDB.head()

#линейная регрессия из коробки
from sklearn.linear_model import LinearRegression

model = LinearRegression()
model.fit(taxiDB.drop(['trip_duration', 'pickup_datetime'], axis=1), taxiDB['trip_duration'])

#коэффициенты и свободный вес
model.coef_.tolist(), model.intercept_

#нахождение обратной матрицы методом Гаусса-Жордана 
def manual_inverse(A):
    A = A.astype(float)
    n = A.shape[0]
    I = np.eye(n)
    AI = np.hstack([A, I])  # augmented matrix

    for i in range(n):
        # Make the diagonal element 1
        AI[i] = AI[i] / AI[i, i]

        # Eliminate other rows
        for j in range(n):
            if i != j:
                AI[j] = AI[j] - AI[j, i] * AI[i]

    return AI[:, n:]

def Manual_inverse_regression(X, Y, fit_intercept=True):
    
    X = np.c_[ X, np.ones(X.shape[0]) ]
    
    return np.dot(np.dot(manual_inverse(np.dot(X.T, X)), X.T), Y)

#нахождение обратной матрицы методом разложения Холецкого
# Linear regression via Cholesky decomposition
def linear_regression_cholesky(X, y):
    #добавление bias
    X_b = np.c_[np.ones((X.shape[0], 1)), X]
    A = X_b.T @ X_b
    b = X_b.T @ y
    
    #разложение Холецкого (built-in in numpy) 
    L = np.linalg.cholesky(A)
    
    #нахождение коэффициентов z в СЛАУ
    z = np.linalg.solve(L, b)
    #обратная подстановка в аналитическое уравнение L^T*theta = z
    theta = np.linalg.solve(L.T, z)
    return theta


#нахождение обратной матрицы методом сингулярного разложения
def linear_regression_svd(X, y):
    #добавление bias
    X_b = np.c_[np.ones((X.shape[0], 1)), X]
    
    #SVD
    U, S, Vt = np.linalg.svd(X_b, full_matrices=False)
    
    #псевдообратная матрица
    S_inv = np.diag(1 / S)
    X_pseudo = Vt.T @ S_inv @ U.T
    theta = X_pseudo @ y
    return theta

print("Cholesky:", linear_regression_cholesky(X_train, Y_train).ravel())
print("SVD:", linear_regression_svd(X_train, Y_train).ravel())

X_train = taxiDB.drop(['trip_duration', 'pickup_datetime'], axis=1).values
Y_train = taxiDB['trip_duration'].values

Manual_inverse_regression(X_train, Y_train)

#реализация аналитического решения уравнения линейной регрессии (built-in methods)
def LinearRegressionByMatrix(X, Y, fit_intercept=True):
    
    X = np.c_[ X, np.ones(X.shape[0]) ]
    
    return np.dot(np.dot(np.linalg.inv(np.dot(X.T, X)), X.T), Y)

#значения ~совпадают со значениями, полученными из линейной регрессии "из коробки"
X_train = taxiDB.drop(['trip_duration', 'pickup_datetime'], axis=1).values
Y_train = taxiDB['trip_duration'].values

LinearRegressionByMatrix(X_train, Y_train)

betas = LinearRegressionByMatrix(X_train, Y_train)

np.dot(np.c_[ X_train, np.ones(X_train.shape[0]) ], betas)
