==> 3_cross_validation.ipynb <==
import pandas as pd

initial_data = pd.read_csv('initial_data.csv', index_col='id')

initial_cols = ['vendor_id', 'passenger_count', 'pickup_longitude',
                'pickup_latitude', 'dropoff_longitude', 'dropoff_latitude',
                'trip_duration']

initial_data = initial_data[initial_cols]

initial_data.head()

#compute MSLE
import numpy as np

initial_data = initial_data.assign(log_trip_duration=np.log1p(initial_data['trip_duration']))
initial_data = initial_data.drop('trip_duration', axis=1)

#Выделим test

from sklearn.model_selection import train_test_split

X = initial_data.drop('log_trip_duration', axis=1)
y = initial_data['log_trip_duration']

X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size=0.2, 
                                                    random_state=42)

#Установим метод кросс-валидации KFold

from sklearn.model_selection import KFold

splitter = KFold(n_splits=20, shuffle=True, random_state=33)

#Воспользуемся методом cross_validate

from sklearn.model_selection import cross_validate
from sklearn.linear_model import LinearRegression

model = LinearRegression()

cv_result = cross_validate(model, X_train, y_train, 
                           scoring='neg_mean_squared_error',
                           cv=splitter, return_train_score=True)

#Возьмем среднее по валидации - качество на кросс-валидации

np.mean(-cv_result['test_score'])

#Финальный замер качества на всей тренировочной выборке
model = LinearRegression()

model.fit(X_train, y_train)

model_predictions = model.predict(X_test)

np.mean(((model_predictions - y_test))**2)

#предобработанные признаки
processed_data = pd.read_csv('processed_data.csv', index_col='id')

processed_data.head()

import numpy as np

processed_data = processed_data.assign(log_trip_duration=np.log1p(processed_data['trip_duration']))
processed_data = processed_data.drop('trip_duration', axis=1)

X_2 = processed_data.drop('log_trip_duration', axis=1)
y_2 = processed_data['log_trip_duration']

test_indexes = X_test.index
train_indexes = X_train.index

X_train_2 = X_2[X_2.index.isin(train_indexes)]
y_train_2 = y_2[y_2.index.isin(train_indexes)]

X_test_2 = X_2[X_2.index.isin(test_indexes)]
y_test_2 = y_2[y_2.index.isin(test_indexes)]

cv_result_2 = cross_validate(model, X_train_2, y_train_2, 
                             scoring='neg_mean_squared_error',
                             cv=splitter, return_train_score=True)

np.mean(-cv_result_2['test_score'])

#Теперь построим модель и замерим качество на тесте. Модель с предобработанными признаками показывает имеет лучшее качество.

model_2 = LinearRegression()

model_2.fit(X_train_2, y_train_2)

model_predictions_2 = model_2.predict(X_test_2)

np.mean(((model_predictions_2 - y_test_2))**2)
