==> 4_regularization.ipynb <==
*Выберите верные утверждения про Lasso и Ridge*: 

1. Ridge-регуляризация более склонна занулять веса модели, нежели Lasso.
2. Lasso-регуляризация более склонна занулять веса модели, нежели Ridge.
3. Ridge и Lasso регуляризация призваны бороться с ситуацией недообучения. 
4. Суть Lasso регуляризации заключается в прибавлении к минимизируемому функционалу суммы модулей обучаемых коэффициентов.
5. Суть Lasso регуляризации заключается в прибавлении к минимизируемому функционалу суммы квадратов обучаемых коэффициентов.



**Ответ: 2, 4)** Оба вида регуляризации, как было показано в лекции, можно свести к поиску условного экстремума на некотором компактном множестве. В случае Ridge регуляризации данное множество оказывается выпуклым (например, в двухмерном пространстве – это круг), ровно как минимизируемый функционал. Тогда шанс того, что линия уровня и граница множества из ограничения соприкоснутся где-то на оси, достаточно маленький. В случае Lasso ситуация обратная (см. Визуальное объяснение в лекции). 
_______________________________________________________

*Выберите верные утверждения про Мультиколлинеарность*: 

1. Мультиколлинеарность гарантирует, что мы получим переобученную модель.
2. Мультиколлинеарность возникает при сильной линейной зависимости в матрице объект-признак.
3. Если в матрице объект-признак присутствуют линейно зависимые признаки, тогда с вероятностью 50% матричную формулу поиска оптимальных коэффициентов регрессии применить нельзя.
4. Если в матрице объект-признак присутствуют линейно зависимые признаки, тогда у минимизируемого функционала оказывается не одна единственная, а уже 2-3 точки минимума.
5. Если в матрице объект-признак присутствуют линейно зависимые признаки, тогда у минимизируемого функционала оказывается бесконечное количество минимумов.
6. Мультиколлинеарность можно полечить, удалив зависимые признаки или используя регуляризацию.




**Ответ: 2, 5, 6)** важно понимать, что данные рассуждения относятся именно к задаче построения линейной регрессии прям как в лекции назад. Для других моделей мультиколлинеарность может, вообще говоря, ни к каким проблемам не приводить.
_______________________________________________________

*Выберите верные утверждения про масштабирование признаков:* 

1. Масштабировать признаки необходимо всегда. Иначе качество на трейне может сильно превышать качество на тесте.
2. Масштабировать признаки рекомендуется всегда. Особенно в задачах, где используется метод градиентного спуска: он становится быстрее.
3. Мы масштабируем признаки, регуляризируя модели, чтобы с одинаковой степенью справедливости учитывать порядок фичей, и, соответственно, чтобы случайно не штрафовать сильно больше те коэффициенты, которые по природе своей должны быть на порядок выше остальных.
4. Подход StandardScaler масштабирует признаки таким образом, что все значения переводятся в промежуток [0; 1].
5. Подход MinMaxScaler масштабирует признаки таким образом, что все значения переводятся в промежуток [0; 1].



**Ответ: 2, 3, 5)** знак производной (положительная она или отрицательная) поясняет нам: возрастает функция или убывает. Минимумы у функции возникают там, где она какое-то время убывала, а потом в моменте стала возрастать. Поэтому такой переход для искомых точек подходит под 1 вариант ответа.

_______________________________________________________

import numpy as np
import pandas as pd
from sklearn.model_selection import KFold, cross_val_score, GridSearchCV
from sklearn.preprocessing import MinMaxScaler
from sklearn.linear_model import LinearRegression, Ridge, Lasso
from sklearn.metrics import make_scorer, mean_squared_error
import warnings

processed_data = pd.read_csv('processed_data.csv', index_col='id')
processed_data = processed_data.assign(log_trip_duration=np.log1p(processed_data['trip_duration']))

processed_data.head()

selector = KFold(n_splits=4, shuffle=True, random_state=33)

model = LinearRegression()

X = processed_data.drop('log_trip_duration', axis=1)
Y = processed_data['log_trip_duration']

cv_result_1 = cross_validate(model, X, Y, 
                             scoring='neg_mean_squared_error',
                             cv=selector, return_train_score=True)

cross_val_error = -np.mean(cv_result_1['test_score'])

print(f"RMSLE на Кросс-валидации: {round(cross_val_error, 3)}")

#dataset with 20 features

new_data = pd.read_csv('enhanced_trip_data.csv', index_col='id')

new_data.head()

features = [col for col in df.columns if col not in ['id', 'trip_duration', 'log_trip_duration']]
X_2 = df[features]
y_2 = df['log_trip_duration']

#MinMaxScaler для стабильности вычислений
scaler = MinMaxScaler()
X_scaled = scaler.fit_transform(X)

#5-fold CV
kf = KFold(n_splits=5, shuffle=True, random_state=42)
#baseline model
lr = LinearRegression()
lr_scores = -cross_val_score(lr, X_scaled, y, cv=kf, scoring=rmse_scorer)

print(f"RMSLE на Кросс-валидации модели с 20 признаками: {np.mean(lr_scores):.4f}")

#Ridge Regression(L2)
ridge = Ridge()
#test different alphas
alphas = np.logspace(-3, 3, 100)
grid_search_ridge = GridSearchCV(ridge, {'alpha': alphas}, scoring=rmse_scorer, cv=kf)
grid_search_ridge.fit(X_scaled, y)

best_ridge_score = -grid_search_ridge.best_score_
print(f"Best Alpha for Ridge: {grid_search_ridge.best_params_['alpha']:.4f}")
print(f"Ridge Regression RMSLE: {best_ridge_score:.4f}")

#Lasso Regression(L1)
lasso = Lasso(max_iter=20000) 
#test different alphas
alphas = np.logspace(-5, -2, 100)
grid_search_lasso = GridSearchCV(lasso, {'alpha': alphas}, scoring=rmse_scorer, cv=kf)
grid_search_lasso.fit(X_scaled, y)

best_lasso_score = -grid_search_lasso.best_score_
best_lasso_model = grid_search_lasso.best_estimator_

print(f"Best Alpha for Lasso: {grid_search_lasso.best_params_['alpha']:.4f}")
print(f"Lasso Regression RMSLE: {best_lasso_score:.4f}")
